+++ src/locs.ts
@@ -1,14 +1,24 @@
 import * as vscode from 'vscode';
-import * as Settings from './settings';
+
+export enum LocsQuickPickType {
+    locsOpenDialog,
+    locsFile,
+    workspaceScan
+}
 
 export interface LocsQuickPickOpen extends vscode.QuickPickItem {
     kind: vscode.QuickPickItemKind.Default;
-    isOpen: true;
+    locsKind: LocsQuickPickType.locsOpenDialog;
+}
+
+export interface LocsQuickPickWorkspaceScan extends vscode.QuickPickItem {
+    kind: vscode.QuickPickItemKind.Default;
+    locsKind: LocsQuickPickType.workspaceScan;
 }
 
 export interface LocsQuickPickFile extends vscode.QuickPickItem {
     kind: vscode.QuickPickItemKind.Default;
-    isOpen: false;
+    locsKind: LocsQuickPickType.locsFile;
     uri: vscode.Uri;
 }
 
@@ -16,7 +26,12 @@ export interface LocsQuickPickSeparator extends vscode.QuickPickItem {
     kind: vscode.QuickPickItemKind.Separator;
 }
 
-export type LocsQuickPickItem = LocsQuickPickOpen | LocsQuickPickFile | LocsQuickPickSeparator;
+export type LocsQuickPickItem = (
+    | LocsQuickPickOpen
+    | LocsQuickPickWorkspaceScan
+    | LocsQuickPickFile
+    | LocsQuickPickSeparator
+);
 
 export function locs(context: vscode.ExtensionContext) {
     return context.workspaceState.get<string>("fpp.locsFile");